[IL-10]DateRange - public class, fix room number type
